moved common types, manager, and middleware implementations to their own crates, initial effort to fix issue #3
