+++ Cargo.toml
@@ -16,8 +16,6 @@ edition = "2018"
 [dependencies]
 anyhow = "1.0"
 async-trait = "0.1"
-bincode = { version = "1.3", optional = true }
-cacache = { version = "9.0", optional = true }
 http = "0.2"
 httpdate = "1.0"
 http-cache-semantics = "0.9"
@@ -31,8 +29,24 @@ task-local-extensions = { version = "0.1", optional = true }
 thiserror = "1.0"
 url = { version = "2.2", features = ["serde"] }
 
+[dependencies.http-cache-types]
+path = "http-cache-types"
+
+[dependencies.http-cache-manager-cacache]
+path = "http-cache-manager-cacache"
+optional = true
+
+[dependencies.http-cache-middleware-reqwest]
+path = "http-cache-middleware-reqwest"
+optional = true
+
+[dependencies.http-cache-middleware-surf]
+path = "http-cache-middleware-surf"
+optional = true
+
 [dev-dependencies]
 async-std = { version = "1.10", features = ["attributes"]}
+cacache = "9.0"
 http-cache-semantics = { version = "0.9", features = ["reqwest"]}
 reqwest = "0.11"
 tokio = { version = "1.15", features = ["macros", "rt-multi-thread"] }
@@ -40,6 +54,19 @@ mockito = "0.30"
 
 [features]
 default = ["manager-cacache"]
-manager-cacache = ["cacache", "bincode"]
-client-reqwest = ["reqwest", "reqwest-middleware", "task-local-extensions"]
-client-surf = ["surf", "http-types"]
+manager-cacache = ["http-cache-manager-cacache", "http-cache-types/manager-cacache"]
+client-reqwest = [
+    "http-cache-types/client-reqwest",
+    "http-cache-middleware-reqwest",
+    "reqwest",
+    "reqwest-middleware",
+    "task-local-extensions"
+]
+client-surf = [
+    "http-cache-types/client-surf",
+    "http-cache-middleware-surf",
+    "surf",
+    "http-types",
+]
+
+[workspace]
