+++ http-cache-manager-cacache/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "http-cache-manager-cacache"
+version = "0.1.0"
+description = "http-cache manager implementation for cacache"
+authors = ["Christian Haynes <06chaynes@gmail.com>", "Kat Marchán <kzm@zkat.tech>"]
+repository = "https://github.com/06chaynes/http-cache.git"
+license = "MIT OR Apache-2.0"
+readme = "../README.md"
+keywords = ["cache", "http", "middleware"]
+categories = [
+    "caching",
+    "web-programming::http-client"
+]
+edition = "2018"
+
+[dependencies]
+async-trait = "0.1"
+bincode = "1.3"
+cacache = "9.0"
+http-cache-semantics = "0.9"
+serde = { version = "1.0", features = ["derive"] }
+url = { version = "2.2", features = ["serde"] }
+
+[dependencies.http-cache-types]
+path = "../http-cache-types"
