+++ http-cache-middleware-reqwest/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "http-cache-middleware-reqwest"
+version = "0.1.0"
+description = "http-cache middleware implementation for reqwest"
+authors = ["Christian Haynes <06chaynes@gmail.com>", "Kat Marchán <kzm@zkat.tech>"]
+repository = "https://github.com/06chaynes/http-cache.git"
+license = "MIT OR Apache-2.0"
+readme = "../README.md"
+keywords = ["cache", "http", "middleware"]
+categories = [
+    "caching",
+    "web-programming::http-client"
+]
+edition = "2018"
+
+[dependencies]
+anyhow = "1.0"
+async-trait = "0.1"
+http = "0.2"
+http-cache-semantics = "0.9"
+reqwest = { version = "0.11", default-features = false}
+reqwest-middleware = "0.1"
+serde = { version = "1.0", features = ["derive"] }
+task-local-extensions = "0.1"
+url = { version = "2.2", features = ["serde"] }
+
+[dependencies.http-cache-types]
+path = "../http-cache-types"
+features = ["client-reqwest"]
