+++ http-cache-middleware-surf/src/lib.rs
@@ -1,0 +1,106 @@
+//! The surf middleware implementation, requires the `client-surf` feature.
+use http_cache_types::{CacheError, HttpResponse, Middleware, Result};
+
+use anyhow::anyhow;
+use std::{
+    collections::HashMap, convert::TryInto, str::FromStr, time::SystemTime,
+};
+
+use http::{header::CACHE_CONTROL, request, request::Parts};
+use http_cache_semantics::{CacheOptions, CachePolicy};
+use http_types::{headers::HeaderValue, Method, Version};
+use surf::{middleware::Next, Client, Request};
+use url::Url;
+
+/// Implements ['Middleware'] for surf
+pub struct SurfMiddleware<'a> {
+    pub req: Request,
+    pub client: Client,
+    pub next: Next<'a>,
+}
+
+#[async_trait::async_trait]
+impl Middleware for SurfMiddleware<'_> {
+    fn is_method_get_head(&self) -> bool {
+        self.req.method() == Method::Get || self.req.method() == Method::Head
+    }
+    fn policy(&self, response: &HttpResponse) -> Result<CachePolicy> {
+        Ok(CachePolicy::new(&self.parts()?, &response.parts()?))
+    }
+    fn policy_with_options(
+        &self,
+        response: &HttpResponse,
+        options: CacheOptions,
+    ) -> Result<CachePolicy> {
+        Ok(CachePolicy::new_options(
+            &self.parts()?,
+            &response.parts()?,
+            SystemTime::now(),
+            options,
+        ))
+    }
+    fn update_headers(&mut self, parts: Parts) -> Result<()> {
+        for header in parts.headers.iter() {
+            let value = match HeaderValue::from_str(header.1.to_str()?) {
+                Ok(v) => v,
+                Err(_e) => return Err(CacheError::BadHeader),
+            };
+            self.req.set_header(header.0.as_str(), value);
+        }
+        Ok(())
+    }
+    fn set_no_cache(&mut self) -> Result<()> {
+        self.req.insert_header(CACHE_CONTROL.as_str(), "no-cache");
+        Ok(())
+    }
+    fn parts(&self) -> Result<Parts> {
+        let mut converted = request::Builder::new()
+            .method(self.req.method().as_ref())
+            .uri(self.req.url().as_str())
+            .body(())?;
+        {
+            let headers = converted.headers_mut();
+            for header in self.req.iter() {
+                headers.insert(
+                    http::header::HeaderName::from_str(header.0.as_str())?,
+                    http::HeaderValue::from_str(header.1.as_str())?,
+                );
+            }
+        }
+        Ok(converted.into_parts().0)
+    }
+    fn url(&self) -> Result<&Url> {
+        Ok(self.req.url())
+    }
+    fn method(&self) -> Result<String> {
+        Ok(self.req.method().as_ref().to_string())
+    }
+    async fn remote_fetch(&mut self) -> Result<HttpResponse> {
+        let url = self.req.url().clone();
+        let mut res =
+            match self.next.run(self.req.clone(), self.client.clone()).await {
+                Ok(r) => r,
+                Err(e) => return Err(CacheError::General(anyhow!(e))),
+            };
+        let mut headers = HashMap::new();
+        for header in res.iter() {
+            headers.insert(
+                header.0.as_str().to_owned(),
+                header.1.as_str().to_owned(),
+            );
+        }
+        let status = res.status().into();
+        let version = res.version().unwrap_or(Version::Http1_1);
+        let body: Vec<u8> = match res.body_bytes().await {
+            Ok(b) => b,
+            Err(e) => return Err(CacheError::General(anyhow!(e))),
+        };
+        Ok(HttpResponse {
+            body,
+            headers,
+            status,
+            url,
+            version: version.try_into()?,
+        })
+    }
+}
