+++ http-cache-types/Cargo.toml
@@ -1,0 +1,35 @@
+[package]
+name = "http-cache-types"
+version = "0.1.0"
+description = "Common types used by http-cache"
+authors = ["Christian Haynes <06chaynes@gmail.com>", "Kat Marchán <kzm@zkat.tech>"]
+repository = "https://github.com/06chaynes/http-cache.git"
+license = "MIT OR Apache-2.0"
+readme = "../README.md"
+keywords = ["cache", "http", "middleware"]
+categories = [
+    "caching",
+    "web-programming::http-client"
+]
+edition = "2018"
+
+[dependencies]
+anyhow = "1.0"
+async-trait = "0.1"
+bincode = { version = "1.3", optional = true }
+cacache = { version = "9.0", optional = true }
+http = "0.2"
+http-cache-semantics = "0.9"
+httpdate = "1.0"
+http-types = { version = "2.12", optional = true }
+miette = "3.3"
+reqwest = { version = "0.11", default-features = false, optional = true }
+reqwest-middleware = { version = "0.1", optional = true }
+serde = { version = "1.0", features = ["derive"] }
+thiserror = "1.0"
+url = { version = "2.2", features = ["serde"] }
+
+[features]
+manager-cacache = ["cacache", "bincode"]
+client-reqwest = ["reqwest", "reqwest-middleware"]
+client-surf = ["http-types"]
