--- src/managers/mod.rs
@@ -1,2 +1,0 @@
-#[cfg(feature = "manager-cacache")]
-pub(crate) mod cacache;
