--- src/middleware/mod.rs
@@ -1,5 +1,0 @@
-#[cfg(feature = "client-surf")]
-pub(crate) mod surf;
-
-#[cfg(feature = "client-reqwest")]
-pub(crate) mod reqwest;
